pyproject.toml
[project]
name = "stfo-colbert"
version = "0.1.0"
description = "Straightforward ColBERT indexing and serving (via PyLate)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pylate>=1.3.4",
]

[project.scripts]
stfo-colbert = "stfo_colbert.cli:main"

[project.optional-dependencies]
# Optional document parsers, install with: uv add "stfo-colbert[docs]"
docs = [
    "pypdf>=4",
    "python-docx>=1",
    "odfpy>=1.4",
]

[tool.setuptools]
exclude-package-data = {"*" = ["example_data/*", "tests/*", "stfo_indexes/*"]}
